renamed packages to fit lowercase naming convention; updated readme to reflect new build path as a result of rename
